+++ portal-impl/test/integration/com/liferay/portlet/InvokerFilterContainerImplTest.java
@@ -29,6 +29,7 @@ import com.liferay.portlet.bundle.invokerfiltercontainerimpl.TestActionFilter;
 import com.liferay.portlet.bundle.invokerfiltercontainerimpl.TestEventFilter;
 import com.liferay.portlet.bundle.invokerfiltercontainerimpl.TestRenderFilter;
 import com.liferay.portlet.bundle.invokerfiltercontainerimpl.TestResourceFilter;
+import com.liferay.portlet.internal.InvokerFilterContainerImpl;
 
 import java.util.List;
